Fixed issues 29 and 35.
Issue 29 was for proper component show/hide functionality.  This was solved by changing Component.show() to Component.visible().  That reflects the method's true intention and removes the ambiguity of naming it the same thing as DeviceScreen.show().  There is no "hide a component" functionality in the LCDUI either (it needs to be done by remove the component from the dialog and adding it back in later).

Fixing issue 35 means that changing a component's value will also recalculate the Dialog's component layout (i.e. invalidate it).  The expensive math involved in invalidation is also now postponed until paint().  Therefore many invalidates can take place between calls to paint() and we only pay the cost once.